python-developer/32-oauth2-i-auth0/szkolenie-01-oauth2-integracja/fastapi-social-auth/requirements.txt
fastapi[standard]
authlib
httpx
PyJWT[crypto]
python-dotenv
itsdangerous
